reorganize search bar
